Update deploy and main files to be suitable and working for MiyooCFW v2 BETA
Add alternative emulator options to simplemenu selectors

Update typo on pico8 rom directory and extension

Update consoles.ini to include PCECD and SEGA CD support

There was also a typo in NES and MAME sections that caused the roms to not load correctly. This should resolve it.

Changes to be committed:
  modified:   simplemenu/output/Bittboy/section_groups/arcades.ini
  modified:   simplemenu/output/Bittboy/section_groups/consoles.ini